+++ src/main/kotlin/com/github/leondevlifelog/gitea/GiteaBundle.kt
@@ -1,0 +1,24 @@
+/*
+ * Copyright (c) 2023. Leon<leondevlifelog@gmail.com>. All rights reserved.
+ * SPDX-License-Identifier: MIT
+ */
+
+package com.github.leondevlifelog.gitea
+
+import com.intellij.DynamicBundle
+import org.jetbrains.annotations.NonNls
+import org.jetbrains.annotations.PropertyKey
+
+@NonNls
+private const val BUNDLE = "messages.GiteaBundle"
+
+object GiteaBundle : DynamicBundle(BUNDLE) {
+
+    @JvmStatic
+    fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) = getMessage(key, *params)
+
+    @Suppress("unused")
+    @JvmStatic
+    fun messagePointer(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) =
+        getLazyMessage(key, *params)
+}