+++ src/main/kotlin/com/github/leondevlifelog/gitea/GiteaConfig.kt
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2023. Leon<leondevlifelog@gmail.com>. All rights reserved.
+ * SPDX-License-Identifier: MIT
+ */
+
+package com.github.leondevlifelog.gitea
+
+/**
+ * global config
+ * @author Leon
+ */
+object GiteaConfig {
+    const val SERVICE_DISPLAY_NAME: String = "Gitea"
+}