+++ src/main/kotlin/com/github/leondevlifelog/gitea/icons/GiteaIcons.kt
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2023. Leon<leondevlifelog@gmail.com>. All rights reserved.
+ * SPDX-License-Identifier: MIT
+ */
+
+package com.github.leondevlifelog.gitea.icons
+
+import com.intellij.openapi.util.IconLoader
+
+/**
+ * all icons in plugin
+ * @author Leon
+ */
+object GiteaIcons {
+    @JvmField
+    val Gitea = IconLoader.getIcon("/icons/gitea.svg", javaClass)
+}