+++ src/main/kotlin/com/github/leondevlifelog/gitea/listeners/GiteaActivationListener.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2023. Leon<leondevlifelog@gmail.com>. All rights reserved.
+ * SPDX-License-Identifier: MIT
+ */
+
+package com.github.leondevlifelog.gitea.listeners
+
+import com.intellij.openapi.application.ApplicationActivationListener
+import com.intellij.openapi.wm.IdeFrame
+/**
+ *
+ * @author Leon
+ */
+internal class GiteaActivationListener : ApplicationActivationListener {
+
+    override fun applicationActivated(ideFrame: IdeFrame) {
+    }
+}