+++ src/main/kotlin/com/github/leondevlifelog/gitea/services/GiteaService.kt
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2023. Leon<leondevlifelog@gmail.com>. All rights reserved.
+ * SPDX-License-Identifier: MIT
+ */
+
+package com.github.leondevlifelog.gitea.services
+
+import com.intellij.openapi.components.Service
+import com.intellij.openapi.diagnostic.thisLogger
+import com.intellij.openapi.project.Project
+import com.github.leondevlifelog.gitea.GiteaBundle
+/**
+ *
+ * @author Leon
+ */
+@Service(Service.Level.PROJECT)
+class GiteaService(project: Project) {
+
+}