+++ src/test/kotlin/com/github/leondevlifelog/gitea/MyPluginTest.kt
@@ -1,4 +1,4 @@
-package com.github.leondevlifelog.giteaideaplugin
+package com.github.leondevlifelog.gitea
 
 import com.intellij.ide.highlighter.XmlFileType
 import com.intellij.openapi.components.service
@@ -6,7 +6,8 @@ import com.intellij.psi.xml.XmlFile
 import com.intellij.testFramework.TestDataPath
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 import com.intellij.util.PsiErrorElementUtil
-import com.github.leondevlifelog.giteaideaplugin.services.MyProjectService
+import com.github.leondevlifelog.gitea.services.GiteaService
+import junit.framework.TestCase
 
 @TestDataPath("\$CONTENT_ROOT/src/test/testData")
 class MyPluginTest : BasePlatformTestCase() {
@@ -30,9 +31,9 @@ class MyPluginTest : BasePlatformTestCase() {
     }
 
     fun testProjectService() {
-        val projectService = project.service<MyProjectService>()
+        val projectService = project.service<GiteaService>()
 
-        assertNotSame(projectService.getRandomNumber(), projectService.getRandomNumber())
+        TestCase.assertSame(GiteaService::class.java, projectService::class.java)
     }
 
     override fun getTestDataPath() = "src/test/testData/rename"